refactor: drop the local REST API, keep webhooks standalone
The desktop-hosted REST server on 127.0.0.1:33443 was superseded by the
Cloud API, which serves the same read endpoints while the app is closed.
Its local API keys only ever authenticated that server, so they go too.

- Delete the axum server and its enable/port settings from plugins/local-api;
  the webhook create path moves into dispatch.rs and webhook fanout no longer
  gates on the server being enabled
- Drop the api_keys table, ops, and types from db-app, and rename the
  webhook-only modules accordingly
- Replace the Developers "Local API" card with a standalone Webhooks section
- Retarget docs: reference/api becomes reference/webhooks, api-cloud lists its
  own endpoints and errors instead of deferring to the local API, and
  /reference/api redirects to the Cloud API page

Webhooks, markdown export automation, and cloud snapshot preparation are
unchanged.
