--- docs/reference/api.mdx
@@ -1,132 +1,0 @@
----
-title: "API reference"
-description: "Endpoints, authentication, and webhooks exposed by the Anarlog local API."
----
-
-The local API is built into the desktop app and available on every plan. Enable it under **Settings → Developers → Local API**. It serves the same read-only meeting data as the CLI and MCP server and can send outbound webhooks.
-
-API reads stay on this device and do not pass through Anarlog's servers. Webhooks are different: Anarlog sends subscribed meeting data to the endpoint you configure.
-
-For hosted access that works while the desktop app is closed, see [Cloud API and connectors](/reference/api-cloud).
-
-The server listens on `http://127.0.0.1:33443` while the desktop app is running. All endpoints except `/health` require an API key created in **Settings → Developers → Local API**.
-
-```bash
-curl -H "Authorization: Bearer anl_..." http://127.0.0.1:33443/v1/meetings
-```
-
-## Authentication
-
-Send the key as a bearer token: `Authorization: Bearer anl_...`. Keys are shown once and stored hashed. Requests without a valid key receive `401`.
-
-Errors returned by Anarlog's endpoint handlers use this envelope:
-
-```json
-{ "error": { "code": "not_found", "message": "meeting 'x' not found" } }
-```
-
-Codes: `unauthorized`, `not_found`, `invalid_request`, `internal_error`.
-
-Malformed JSON can return a plain-text `400` before the endpoint handler runs.
-
-## Endpoints
-
-### `GET /health`
-
-Unauthenticated liveness check. Returns `{ "status": "ok", "version": "..." }`.
-
-### `GET /v1/meetings`
-
-| Parameter | Type | Default | Notes |
-| --- | --- | --- | --- |
-| `query` | string | — | Case-insensitive title or meeting ID substring. |
-| `series_id` | string | — | Exact recurring series ID. |
-| `limit` | integer | `20` | Clamped to `1..200`. |
-| `offset` | integer | `0` | Number of results to skip. |
-
-Returns `{ "meetings": [...], "pagination": {...} }`. Pagination contains `offset`, `limit`, `returned`, `total`, and `next_offset`; a missing `next_offset` means the page reached the end.
-
-### `GET /v1/meetings/{meeting_id}`
-
-Returns meeting metadata, the canonical note, summaries, participants, and action items. Note and summary bodies are markdown.
-
-### `GET /v1/meetings/{meeting_id}/transcript`
-
-| Parameter | Type | Default | Notes |
-| --- | --- | --- | --- |
-| `offset` | integer | `0` | Word offset. |
-| `limit` | integer | `200` | Clamped to `1..500`. |
-
-Returns `meeting_id`, joined `text`, the raw `words`, and `pagination`.
-
-### `GET /v1/meetings/{meeting_id}/history`
-
-Meetings in the same recurring series, newest first. Accepts `limit` (default 20, clamped to `1..200`) and `offset`.
-
-### `GET /v1/meetings/{meeting_id}/export`
-
-Full meeting export including transcripts. `format=json` (default) returns the structured export; `format=markdown` returns a rendered markdown document.
-
-### `GET /v1/webhooks` · `POST /v1/webhooks` · `DELETE /v1/webhooks/{id}` · `POST /v1/webhooks/{id}/test`
-
-Manage webhook endpoints. Create with:
-
-```json
-{ "url": "https://example.com/hooks", "events": ["note.enhanced"] }
-```
-
-An empty `events` array subscribes to all events. A successful create returns `201` and includes the signing `secret` (`whsec_...`) exactly once. List responses omit the secret. Delete returns `204`; test returns `delivered` and `status` for one delivery attempt.
-
-## Webhooks
-
-Anarlog delivers events as JSON `POST` requests while the desktop app is running. Meeting events include the note, summaries, participants, action items, and full transcript text, so only use an endpoint you trust.
-
-| Event | Fires when |
-| --- | --- |
-| `meeting.completed` | A recording session finishes. |
-| `note.enhanced` | An AI summary is generated for a meeting. |
-| `webhook.test` | You trigger a test delivery. |
-
-The payload envelope is:
-
-```json
-{
-  "id": "evt_...",
-  "event": "note.enhanced",
-  "created_at": "2026-07-28T09:00:00.000Z",
-  "data": {
-    "meeting": { "id": "...", "title": "...", "note": {...}, "summaries": [...], "participants": [...], "action_items": [...] },
-    "transcript_text": "..."
-  }
-}
-```
-
-Meeting event deliveries time out after 10 seconds. Failed attempts retry after 5 and 30 seconds. A `webhook.test` request makes one attempt. Each request includes:
-
-| Header | Content |
-| --- | --- |
-| `x-anarlog-event` | Event name. |
-| `x-anarlog-delivery` | Unique delivery ID. |
-| `x-anarlog-timestamp` | Unix timestamp (seconds). |
-| `x-anarlog-signature` | `sha256=<hex>` — HMAC-SHA256 of the raw request body using your endpoint's `whsec_...` secret. |
-
-Verify signatures by computing the HMAC over the exact raw body:
-
-```js
-import crypto from "node:crypto";
-
-const expected = `sha256=${
-  crypto.createHmac("sha256", secret).update(rawBody).digest("hex")
-}`;
-const signature = request.headers["x-anarlog-signature"];
-const received = typeof signature === "string" ? signature : "";
-const valid =
-  Buffer.byteLength(received) === Buffer.byteLength(expected) &&
-  crypto.timingSafeEqual(Buffer.from(received), Buffer.from(expected));
-```
-
-Verify the raw body before parsing it. For replay protection, reject stale signed `created_at` values and store each payload `id` so retries are processed once.
-
-## Limits
-
-The server binds to `127.0.0.1` only and runs while the desktop app is open. There are no request rate limits. List endpoints cap `limit` at 200, transcript pages at 500 words, and one installation can configure at most 64 webhook endpoints.
