+++ plugins/local-api/Cargo.toml
@@ -4,7 +4,7 @@ version = "0.1.0"
 edition = "2024"
 exclude = ["/js", "/node_modules"]
 links = "tauri-plugin-local-api"
-description = "Local HTTP API for Pro users: API-key-authenticated meeting endpoints and outbound webhooks"
+description = "Outbound meeting webhooks, markdown export, and cloud snapshot preparation"
 
 [build-dependencies]
 tauri-plugin = { workspace = true, features = ["build"] }
@@ -18,7 +18,6 @@ anlg-agent-access = { workspace = true }
 anlg-db-app = { workspace = true }
 tauri-plugin-db = { workspace = true }
 
-axum = { workspace = true }
 bytes = { workspace = true }
 chrono = { workspace = true }
 hmac = { workspace = true }
@@ -30,7 +29,6 @@ specta = { workspace = true }
 sqlx = { workspace = true, features = ["runtime-tokio", "sqlite"] }
 tauri = { workspace = true, features = ["test", "macos-private-api"] }
 tauri-specta = { workspace = true, features = ["derive", "typescript"] }
-thiserror = { workspace = true }
-tokio = { workspace = true, features = ["macros", "net", "sync", "rt", "time"] }
+tokio = { workspace = true, features = ["macros", "sync", "rt", "time"] }
 tracing = { workspace = true }
 uuid = { workspace = true, features = ["v4"] }
