+++ plugins/local-api/build.rs
@@ -1,12 +1,8 @@
 const COMMANDS: &[&str] = &[
-    "get_status",
-    "set_enabled",
-    "list_api_keys",
-    "create_api_key",
-    "revoke_api_key",
     "list_webhooks",
     "create_webhook",
     "delete_webhook",
+    "set_webhook_active",
     "test_webhook",
     "dispatch_event",
     "export_meeting_markdown",
