--- plugins/local-api/permissions/autogenerated/commands/create_api_key.toml
@@ -1,13 +1,0 @@
-# Automatically generated - DO NOT EDIT!
-
-"$schema" = "../../schemas/schema.json"
-
-[[permission]]
-identifier = "allow-create-api-key"
-description = "Enables the create_api_key command without any pre-configured scope."
-commands.allow = ["create_api_key"]
-
-[[permission]]
-identifier = "deny-create-api-key"
-description = "Denies the create_api_key command without any pre-configured scope."
-commands.deny = ["create_api_key"]
