--- plugins/local-api/permissions/autogenerated/commands/get_status.toml
@@ -1,13 +1,0 @@
-# Automatically generated - DO NOT EDIT!
-
-"$schema" = "../../schemas/schema.json"
-
-[[permission]]
-identifier = "allow-get-status"
-description = "Enables the get_status command without any pre-configured scope."
-commands.allow = ["get_status"]
-
-[[permission]]
-identifier = "deny-get-status"
-description = "Denies the get_status command without any pre-configured scope."
-commands.deny = ["get_status"]
