--- plugins/local-api/permissions/autogenerated/commands/revoke_api_key.toml
@@ -1,13 +1,0 @@
-# Automatically generated - DO NOT EDIT!
-
-"$schema" = "../../schemas/schema.json"
-
-[[permission]]
-identifier = "allow-revoke-api-key"
-description = "Enables the revoke_api_key command without any pre-configured scope."
-commands.allow = ["revoke_api_key"]
-
-[[permission]]
-identifier = "deny-revoke-api-key"
-description = "Denies the revoke_api_key command without any pre-configured scope."
-commands.deny = ["revoke_api_key"]
