--- plugins/local-api/permissions/autogenerated/commands/set_enabled.toml
@@ -1,13 +1,0 @@
-# Automatically generated - DO NOT EDIT!
-
-"$schema" = "../../schemas/schema.json"
-
-[[permission]]
-identifier = "allow-set-enabled"
-description = "Enables the set_enabled command without any pre-configured scope."
-commands.allow = ["set_enabled"]
-
-[[permission]]
-identifier = "deny-set-enabled"
-description = "Denies the set_enabled command without any pre-configured scope."
-commands.deny = ["set_enabled"]
