+++ plugins/local-api/permissions/schemas/schema.json
@@ -295,18 +295,6 @@
       "type": "string",
       "oneOf": [
         {
-          "description": "Enables the create_api_key command without any pre-configured scope.",
-          "type": "string",
-          "const": "allow-create-api-key",
-          "markdownDescription": "Enables the create_api_key command without any pre-configured scope."
-        },
-        {
-          "description": "Denies the create_api_key command without any pre-configured scope.",
-          "type": "string",
-          "const": "deny-create-api-key",
-          "markdownDescription": "Denies the create_api_key command without any pre-configured scope."
-        },
-        {
           "description": "Enables the create_webhook command without any pre-configured scope.",
           "type": "string",
           "const": "allow-create-webhook",
@@ -367,30 +355,6 @@
           "markdownDescription": "Denies the get_cloud_snapshot command without any pre-configured scope."
         },
         {
-          "description": "Enables the get_status command without any pre-configured scope.",
-          "type": "string",
-          "const": "allow-get-status",
-          "markdownDescription": "Enables the get_status command without any pre-configured scope."
-        },
-        {
-          "description": "Denies the get_status command without any pre-configured scope.",
-          "type": "string",
-          "const": "deny-get-status",
-          "markdownDescription": "Denies the get_status command without any pre-configured scope."
-        },
-        {
-          "description": "Enables the list_api_keys command without any pre-configured scope.",
-          "type": "string",
-          "const": "allow-list-api-keys",
-          "markdownDescription": "Enables the list_api_keys command without any pre-configured scope."
-        },
-        {
-          "description": "Denies the list_api_keys command without any pre-configured scope.",
-          "type": "string",
-          "const": "deny-list-api-keys",
-          "markdownDescription": "Denies the list_api_keys command without any pre-configured scope."
-        },
-        {
           "description": "Enables the list_cloud_snapshot_ids command without any pre-configured scope.",
           "type": "string",
           "const": "allow-list-cloud-snapshot-ids",
@@ -415,28 +379,16 @@
           "markdownDescription": "Denies the list_webhooks command without any pre-configured scope."
         },
         {
-          "description": "Enables the revoke_api_key command without any pre-configured scope.",
+          "description": "Enables the set_webhook_active command without any pre-configured scope.",
           "type": "string",
-          "const": "allow-revoke-api-key",
-          "markdownDescription": "Enables the revoke_api_key command without any pre-configured scope."
+          "const": "allow-set-webhook-active",
+          "markdownDescription": "Enables the set_webhook_active command without any pre-configured scope."
         },
         {
-          "description": "Denies the revoke_api_key command without any pre-configured scope.",
+          "description": "Denies the set_webhook_active command without any pre-configured scope.",
           "type": "string",
-          "const": "deny-revoke-api-key",
-          "markdownDescription": "Denies the revoke_api_key command without any pre-configured scope."
-        },
-        {
-          "description": "Enables the set_enabled command without any pre-configured scope.",
-          "type": "string",
-          "const": "allow-set-enabled",
-          "markdownDescription": "Enables the set_enabled command without any pre-configured scope."
-        },
-        {
-          "description": "Denies the set_enabled command without any pre-configured scope.",
-          "type": "string",
-          "const": "deny-set-enabled",
-          "markdownDescription": "Denies the set_enabled command without any pre-configured scope."
+          "const": "deny-set-webhook-active",
+          "markdownDescription": "Denies the set_webhook_active command without any pre-configured scope."
         },
         {
           "description": "Enables the test_webhook command without any pre-configured scope.",
@@ -451,10 +403,10 @@
           "markdownDescription": "Denies the test_webhook command without any pre-configured scope."
         },
         {
-          "description": "Default permissions for the plugin\n#### This default permission set includes:\n\n- `allow-get-status`\n- `allow-set-enabled`\n- `allow-list-api-keys`\n- `allow-create-api-key`\n- `allow-revoke-api-key`\n- `allow-list-webhooks`\n- `allow-create-webhook`\n- `allow-delete-webhook`\n- `allow-test-webhook`\n- `allow-dispatch-event`\n- `allow-export-meeting-markdown`\n- `allow-get-cloud-snapshot`\n- `allow-list-cloud-snapshot-ids`",
+          "description": "Default permissions for the plugin\n#### This default permission set includes:\n\n- `allow-list-webhooks`\n- `allow-create-webhook`\n- `allow-delete-webhook`\n- `allow-set-webhook-active`\n- `allow-test-webhook`\n- `allow-dispatch-event`\n- `allow-export-meeting-markdown`\n- `allow-get-cloud-snapshot`\n- `allow-list-cloud-snapshot-ids`",
           "type": "string",
           "const": "default",
-          "markdownDescription": "Default permissions for the plugin\n#### This default permission set includes:\n\n- `allow-get-status`\n- `allow-set-enabled`\n- `allow-list-api-keys`\n- `allow-create-api-key`\n- `allow-revoke-api-key`\n- `allow-list-webhooks`\n- `allow-create-webhook`\n- `allow-delete-webhook`\n- `allow-test-webhook`\n- `allow-dispatch-event`\n- `allow-export-meeting-markdown`\n- `allow-get-cloud-snapshot`\n- `allow-list-cloud-snapshot-ids`"
+          "markdownDescription": "Default permissions for the plugin\n#### This default permission set includes:\n\n- `allow-list-webhooks`\n- `allow-create-webhook`\n- `allow-delete-webhook`\n- `allow-set-webhook-active`\n- `allow-test-webhook`\n- `allow-dispatch-event`\n- `allow-export-meeting-markdown`\n- `allow-get-cloud-snapshot`\n- `allow-list-cloud-snapshot-ids`"
         }
       ]
     }
