+++ plugins/local-api/src/commands.rs
@@ -1,9 +1,6 @@
 use tauri::Manager;
 
-use crate::{
-    ApiKeyInfo, CreatedApiKey, CreatedWebhook, LocalApiStatus, PluginState, WebhookDelivery,
-    WebhookInfo, dispatch, server, settings,
-};
+use crate::{CreatedWebhook, WebhookDelivery, WebhookInfo, dispatch};
 
 const MAX_CLOUD_SNAPSHOT_BYTES: usize = 2 * 1024 * 1024;
 
@@ -11,131 +8,6 @@ fn pool<R: tauri::Runtime>(app: &tauri::AppHandle<R>) -> Result<sqlx::SqlitePool
     app.try_state::<tauri_plugin_db::ManagedState>()
         .map(|state| state.pool().clone())
         .ok_or_else(|| "database is not ready yet".to_string())
-}
-
-#[tauri::command]
-#[specta::specta]
-pub async fn get_status<R: tauri::Runtime>(
-    app: tauri::AppHandle<R>,
-) -> Result<LocalApiStatus, String> {
-    let state = app.state::<PluginState>();
-    let running_port = state.server.lock().await.as_ref().map(|handle| handle.port);
-    let settings = match pool(&app) {
-        Ok(pool) => settings::load(&pool).await.unwrap_or_default(),
-        Err(_) => settings::LocalApiSettings::default(),
-    };
-    Ok(LocalApiStatus {
-        enabled: settings.enabled,
-        port: running_port.unwrap_or(settings.port),
-        running: running_port.is_some(),
-    })
-}
-
-#[tauri::command]
-#[specta::specta]
-pub async fn set_enabled<R: tauri::Runtime>(
-    app: tauri::AppHandle<R>,
-    enabled: bool,
-) -> Result<LocalApiStatus, String> {
-    let pool = pool(&app)?;
-    let state = app.state::<PluginState>();
-    let mut server = state.server.lock().await;
-    let mut settings = settings::load(&pool).await.map_err(|e| e.to_string())?;
-    if enabled {
-        if server.is_none() {
-            let handle = match server::start(pool.clone(), settings.port).await {
-                Ok(handle) => handle,
-                Err(error) => {
-                    if settings.enabled {
-                        settings.enabled = false;
-                        settings::store(&pool, &settings)
-                            .await
-                            .map_err(|rollback| {
-                                format!(
-                                    "{error}; could not clear enabled setting after bind failure: {rollback}"
-                                )
-                            })?;
-                    }
-                    return Err(error);
-                }
-            };
-            settings.enabled = true;
-            if let Err(error) = settings::store(&pool, &settings).await {
-                handle.shutdown().await;
-                return Err(error.to_string());
-            }
-            *server = Some(handle);
-        } else {
-            settings.enabled = true;
-            settings::store(&pool, &settings)
-                .await
-                .map_err(|e| e.to_string())?;
-        }
-    } else {
-        settings.enabled = false;
-        settings::store(&pool, &settings)
-            .await
-            .map_err(|e| e.to_string())?;
-        if let Some(handle) = server.take() {
-            handle.shutdown().await;
-        }
-    }
-
-    let running_port = server.as_ref().map(|handle| handle.port);
-    Ok(LocalApiStatus {
-        enabled: settings.enabled,
-        port: running_port.unwrap_or(settings.port),
-        running: running_port.is_some(),
-    })
-}
-
-#[tauri::command]
-#[specta::specta]
-pub async fn list_api_keys<R: tauri::Runtime>(
-    app: tauri::AppHandle<R>,
-) -> Result<Vec<ApiKeyInfo>, String> {
-    let pool = pool(&app)?;
-    Ok(anlg_db_app::list_api_keys(&pool)
-        .await
-        .map_err(|e| e.to_string())?
-        .into_iter()
-        .map(ApiKeyInfo::from)
-        .collect())
-}
-
-#[tauri::command]
-#[specta::specta]
-pub async fn create_api_key<R: tauri::Runtime>(
-    app: tauri::AppHandle<R>,
-    name: String,
-) -> Result<CreatedApiKey, String> {
-    let pool = pool(&app)?;
-    let generated = anlg_db_app::generate_api_key();
-    let row = anlg_db_app::insert_api_key(
-        &pool,
-        &uuid::Uuid::new_v4().to_string(),
-        name.trim(),
-        &generated.key_hash,
-        &generated.key_prefix,
-    )
-    .await
-    .map_err(|e| e.to_string())?;
-    Ok(CreatedApiKey {
-        info: ApiKeyInfo::from(row),
-        key: generated.key,
-    })
-}
-
-#[tauri::command]
-#[specta::specta]
-pub async fn revoke_api_key<R: tauri::Runtime>(
-    app: tauri::AppHandle<R>,
-    id: String,
-) -> Result<bool, String> {
-    let pool = pool(&app)?;
-    anlg_db_app::revoke_api_key(&pool, &id)
-        .await
-        .map_err(|e| e.to_string())
 }
 
 #[tauri::command]
@@ -160,7 +32,7 @@ pub async fn create_webhook<R: tauri::Runtime>(
     events: Vec<String>,
 ) -> Result<CreatedWebhook, String> {
     let pool = pool(&app)?;
-    server::create_endpoint(&pool, &url, &events).await
+    dispatch::create_endpoint(&pool, &url, &events).await
 }
 
 #[tauri::command]
@@ -173,6 +45,21 @@ pub async fn delete_webhook<R: tauri::Runtime>(
     anlg_db_app::delete_webhook_endpoint(&pool, &id)
         .await
         .map_err(|e| e.to_string())
+}
+
+#[tauri::command]
+#[specta::specta]
+pub async fn set_webhook_active<R: tauri::Runtime>(
+    app: tauri::AppHandle<R>,
+    id: String,
+    active: bool,
+) -> Result<WebhookInfo, String> {
+    let pool = pool(&app)?;
+    anlg_db_app::set_webhook_endpoint_active(&pool, &id, active)
+        .await
+        .map_err(|e| e.to_string())?
+        .map(WebhookInfo::from)
+        .ok_or_else(|| format!("webhook '{id}' not found"))
 }
 
 #[tauri::command]
