+++ plugins/local-api/src/types.rs
@@ -2,44 +2,6 @@ use serde::{Deserialize, Serialize};
 
 #[derive(Debug, Clone, PartialEq, Eq, Serialize, Deserialize, specta::Type)]
 #[serde(rename_all = "snake_case")]
-pub struct LocalApiStatus {
-    pub enabled: bool,
-    pub port: u16,
-    pub running: bool,
-}
-
-#[derive(Debug, Clone, PartialEq, Eq, Serialize, Deserialize, specta::Type)]
-#[serde(rename_all = "snake_case")]
-pub struct ApiKeyInfo {
-    pub id: String,
-    pub name: String,
-    pub key_prefix: String,
-    pub created_at: String,
-    pub last_used_at: Option<String>,
-}
-
-impl From<anlg_db_app::ApiKeyRow> for ApiKeyInfo {
-    fn from(row: anlg_db_app::ApiKeyRow) -> Self {
-        Self {
-            id: row.id,
-            name: row.name,
-            key_prefix: row.key_prefix,
-            created_at: row.created_at,
-            last_used_at: row.last_used_at,
-        }
-    }
-}
-
-#[derive(Debug, Clone, PartialEq, Eq, Serialize, Deserialize, specta::Type)]
-#[serde(rename_all = "snake_case")]
-pub struct CreatedApiKey {
-    #[serde(flatten)]
-    pub info: ApiKeyInfo,
-    pub key: String,
-}
-
-#[derive(Debug, Clone, PartialEq, Eq, Serialize, Deserialize, specta::Type)]
-#[serde(rename_all = "snake_case")]
 pub struct WebhookInfo {
     pub id: String,
     pub url: String,
